# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Task Management",
        "description": "Sample Project to Demonstrate Laravel API Implementation with Passport.",
        "contact": {
            "email": "route.imdrashedul@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a list of paginated projects",
                "description": "Returns a list of projects with pagination",
                "operationId": "229d90e5cb689f53753846354ec9b7d5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of projects per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of paginated projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessProjectPaginated"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JwtToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "64e2c90ec966de73f1ae3361106288bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessProjectCreated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorProjectValidation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JwtToken": []
                    }
                ]
            }
        },
        "/api/v1/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a single project by its ID",
                "operationId": "4e1905b8e174199605b555cac5922167",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of a single project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessProjectFetch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorProjectNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JwtToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "operationId": "716bae898aeaa6b32146f63b4ecd1fbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessProjectUpdated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorProjectValidation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorProjectNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JwtToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "operationId": "c80ebc592a8f9d46a038d50a7da42008",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessProjectDeleted"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorProjectNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JwtToken": []
                    }
                ]
            }
        },
        "/api/v1/projects/{id}/tasks": {
            "get": {
                "tags": [
                    "Projects",
                    "Tasks"
                ],
                "summary": "Get tasks for a project",
                "operationId": "fa83b1b38a71ff5436eb12fe03ce7dbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of projects per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks for the project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessTasksPaginated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorProjectNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JwtToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorUnauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
...
            "SuccessProjectPaginated": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "pagination": {
                        "properties": {
                            "total": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "current_page": {
                                "type": "integer"
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": "integer"
                            },
                            "to": {
                                "type": "integer"
                            },
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            },
                            "next": {
                                "type": "string"
                            },
                            "prev": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessProjectFetch": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Project fetched"
                    },
                    "success": {
                        "properties": {
                            "project": {
                                "$ref": "#/components/schemas/Project"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorProjectNotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Project not found"
                    },
                    "errors": {
                        "properties": {
                            "project": {
                                "type": "object",
                                "example": "Bad request or project not exists"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ProjectRequest": {
                "required": [
                    "title",
                    "status",
                    "priority",
                    "description",
                    "deadline"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Sample Project"
                    },
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "priority": {
                        "type": "integer",
                        "example": 2
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-05 02:59:00"
                    },
                    "description": {
                        "type": "string",
                        "example": "Project description"
                    }
                },
                "type": "object"
            },
            "SuccessProjectCreated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Project created"
                    },
                    "success": {
                        "properties": {
                            "project": {
                                "$ref": "#/components/schemas/Project"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorProjectValidation": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Project title is required (and 3 more errors)"
                    },
                    "errors": {
                        "properties": {
                            "title": {
                                "description": "Array of error messages related to title field",
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Project title is required."
                                }
                            },
                            "status": {
                                "description": "Array of error messages related to status field",
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Project status is required."
                                }
                            },
                            "priority": {
                                "description": "Array of error messages related to priority field",
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Project priority is required."
                                }
                            },
                            "description": {
                                "description": "Array of error messages related to description field",
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Project description is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessProjectUpdated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Project updated"
                    },
                    "success": {
                        "properties": {
                            "project": {
                                "properties": {
                                    "old": {
                                        "$ref": "#/components/schemas/Project"
                                    },
                                    "new": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessProjectDeleted": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Project deleted"
                    },
                    "success": {
                        "properties": {
                            "project": {
                                "properties": {
                                    "deleted": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JwtToken": {
                "type": "http",
                "description": "Enter your JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}